Fix the logic flaw with %ignore
The behaviour of the %ignore flag does not match the documentation. At
the moment, if %ignore is set and an existing value is found, a
multivalued attribute is created. We need to invert the sense of %ignore
in the code to correct this (and make it match the documentation).

This fixes #1